refactor(test): split noop_gate tests into 3 files for LOC compliance
Split test_noop_gate_unit.py (668 lines) into 3 files to meet CI 400-line limit:
- test_noop_gate_unit.py: Core state change tests (303 lines, 11 tests)
- test_noop_gate_ref_check.py: Ref-check tests (143 lines, 4 tests)
- test_noop_gate_transition_count.py: Transition count tests (263 lines, 8 tests)

All 23 tests pass, mypy clean, ruff clean.

Related: #822
