+++ tests/vibe3/execution/test_noop_gate_unit.py
@@ -301,368 +301,3 @@ class TestApplyUnifiedNoopGate:
         # Gate should skip, not block
         mock_block.assert_not_called()
         store.add_event.assert_not_called()
-
-    def test_blocks_when_required_ref_missing_even_if_state_changed(self) -> None:
-        """Gate blocks when required ref is missing, even if state changed."""
-        store = _make_mock_store()
-
-        with (
-            patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
-            patch(
-                "vibe3.services.issue_failure_service.block_planner_noop_issue"
-            ) as mock_block,
-        ):
-            mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
-                "state/in-progress"
-            )
-            apply_unified_noop_gate(
-                store=store,
-                issue_number=42,
-                branch="task/issue-42",
-                actor="agent:plan",
-                role="planner",
-                before_state_label="state/plan",
-                required_ref_key="plan_ref",
-                flow_state={},  # plan_ref missing
-            )
-
-        mock_block.assert_called_once()
-        call_kwargs = mock_block.call_args[1]
-        assert "required ref missing" in call_kwargs["reason"]
-        assert "plan_ref" in call_kwargs["reason"]
-        store.add_event.assert_called_once()
-        event_args = store.add_event.call_args
-        assert event_args[0][1] == "required_ref_missing"
-
-    def test_passes_when_ref_present_and_state_changed(self) -> None:
-        """Gate passes when ref is present and state changed."""
-        store = _make_mock_store()
-
-        with (
-            patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
-            patch(
-                "vibe3.services.issue_failure_service.block_executor_noop_issue"
-            ) as mock_block,
-        ):
-            mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
-                "state/review"
-            )
-            apply_unified_noop_gate(
-                store=store,
-                issue_number=99,
-                branch="task/issue-99",
-                actor="agent:run",
-                role="executor",
-                before_state_label="state/run",
-                required_ref_key="report_ref",
-                flow_state={"report_ref": "path/to/report.md"},
-            )
-
-        mock_block.assert_not_called()
-        store.add_event.assert_called_once()
-        event_args = store.add_event.call_args
-        assert event_args[0][1] == "state_transitioned"
-
-    def test_blocks_when_ref_present_but_state_unchanged(self) -> None:
-        """Gate blocks when ref is present but state unchanged."""
-        store = _make_mock_store()
-
-        with (
-            patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
-            patch(
-                "vibe3.services.issue_failure_service.block_reviewer_noop_issue"
-            ) as mock_block,
-        ):
-            mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
-                "state/review"
-            )
-            apply_unified_noop_gate(
-                store=store,
-                issue_number=55,
-                branch="task/issue-55",
-                actor="agent:review",
-                role="reviewer",
-                before_state_label="state/review",
-                required_ref_key="audit_ref",
-                flow_state={"audit_ref": "path/to/audit.md"},
-            )
-
-        mock_block.assert_called_once()
-        call_kwargs = mock_block.call_args[1]
-        assert "state unchanged" in call_kwargs["reason"]
-        store.add_event.assert_called_once()
-        event_args = store.add_event.call_args
-        assert event_args[0][1] == "state_unchanged"
-
-    def test_manager_skips_ref_check_state_changed_passes(self) -> None:
-        """Manager role skips ref check; only state change matters."""
-        store = _make_mock_store()
-
-        with (
-            patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
-            patch(
-                "vibe3.services.issue_failure_service.block_manager_noop_issue"
-            ) as mock_block,
-        ):
-            mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
-                "state/handoff"
-            )
-            apply_unified_noop_gate(
-                store=store,
-                issue_number=42,
-                branch="task/issue-42",
-                actor="agent:manager",
-                role="manager",
-                before_state_label="state/ready",
-                required_ref_key=None,  # manager has no required ref
-                flow_state=None,
-            )
-
-        mock_block.assert_not_called()
-        store.add_event.assert_called_once()
-        event_args = store.add_event.call_args
-        assert event_args[0][1] == "state_transitioned"
-
-
-class TestTransitionCount:
-    """Tests for transition_count increment and limit checking."""
-
-    def test_transition_count_incremented_on_state_change(self) -> None:
-        """transition_count is incremented when state changes."""
-        store = _make_mock_store()
-        flow_state = {"transition_count": 3}
-
-        with (
-            patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
-            patch(
-                "vibe3.services.issue_failure_service.block_planner_noop_issue"
-            ) as mock_block,
-        ):
-            mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
-                "state/ready"
-            )
-            apply_unified_noop_gate(
-                store=store,
-                issue_number=42,
-                branch="task/issue-42",
-                actor="agent:plan",
-                role="planner",
-                before_state_label="state/plan",
-                flow_state=flow_state,
-            )
-
-        mock_block.assert_not_called()
-        assert flow_state["transition_count"] == 4
-        store.add_event.assert_called_once()
-        event_args = store.add_event.call_args
-        assert event_args[0][1] == "state_transitioned"
-
-    def test_transition_count_not_incremented_on_state_unchanged(self) -> None:
-        """transition_count is NOT incremented when state is unchanged."""
-        store = _make_mock_store()
-        flow_state = {"transition_count": 3}
-
-        with (
-            patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
-            patch(
-                "vibe3.services.issue_failure_service.block_planner_noop_issue"
-            ) as mock_block,
-        ):
-            mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
-                "state/plan"
-            )
-            apply_unified_noop_gate(
-                store=store,
-                issue_number=42,
-                branch="task/issue-42",
-                actor="agent:plan",
-                role="planner",
-                before_state_label="state/plan",
-                flow_state=flow_state,
-            )
-
-        mock_block.assert_called_once()
-        assert flow_state["transition_count"] == 3  # Unchanged
-        store.add_event.assert_called_once()
-        event_args = store.add_event.call_args
-        assert event_args[0][1] == "state_unchanged"
-
-    def test_transition_count_hard_limit_blocks(self) -> None:
-        """Gate blocks when transition_count reaches hard limit (20)."""
-        store = _make_mock_store()
-        flow_state = {"transition_count": 19}
-
-        with (
-            patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
-            patch(
-                "vibe3.services.issue_failure_service.block_executor_noop_issue"
-            ) as mock_block,
-        ):
-            mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
-                "state/review"
-            )
-            apply_unified_noop_gate(
-                store=store,
-                issue_number=99,
-                branch="task/issue-99",
-                actor="agent:run",
-                role="executor",
-                before_state_label="state/run",
-                flow_state=flow_state,
-            )
-
-        mock_block.assert_called_once()
-        call_kwargs = mock_block.call_args[1]
-        assert "transition count exceeded" in call_kwargs["reason"]
-        store.add_event.assert_called_once()
-        event_args = store.add_event.call_args
-        assert event_args[0][1] == "transition_count_exceeded"
-
-    def test_transition_count_soft_limit_warns_no_block(self) -> None:
-        """Soft limit (10) only warns, does not block."""
-        store = _make_mock_store()
-        flow_state = {"transition_count": 9}
-
-        with (
-            patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
-            patch(
-                "vibe3.services.issue_failure_service.block_executor_noop_issue"
-            ) as mock_block,
-        ):
-            mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
-                "state/review"
-            )
-            apply_unified_noop_gate(
-                store=store,
-                issue_number=99,
-                branch="task/issue-99",
-                actor="agent:run",
-                role="executor",
-                before_state_label="state/run",
-                flow_state=flow_state,
-            )
-
-        # Soft limit does NOT block
-        mock_block.assert_not_called()
-        assert flow_state["transition_count"] == 10
-        store.add_event.assert_called_once()
-        event_args = store.add_event.call_args
-        assert event_args[0][1] == "state_transitioned"
-
-    def test_transition_count_below_soft_limit_no_warning(self) -> None:
-        """No warning when transition_count below soft limit."""
-        store = _make_mock_store()
-        flow_state = {"transition_count": 3}
-
-        with (
-            patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
-            patch(
-                "vibe3.services.issue_failure_service.block_executor_noop_issue"
-            ) as mock_block,
-        ):
-            mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
-                "state/review"
-            )
-            apply_unified_noop_gate(
-                store=store,
-                issue_number=99,
-                branch="task/issue-99",
-                actor="agent:run",
-                role="executor",
-                before_state_label="state/run",
-                flow_state=flow_state,
-            )
-
-        mock_block.assert_not_called()
-        assert flow_state["transition_count"] == 4
-        store.add_event.assert_called_once()
-        event_args = store.add_event.call_args
-        assert event_args[0][1] == "state_transitioned"
-
-    def test_transition_count_none_flow_state_skips_check(self) -> None:
-        """transition_count logic is skipped when flow_state is None."""
-        store = _make_mock_store()
-
-        with (
-            patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
-            patch(
-                "vibe3.services.issue_failure_service.block_executor_noop_issue"
-            ) as mock_block,
-        ):
-            mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
-                "state/review"
-            )
-            apply_unified_noop_gate(
-                store=store,
-                issue_number=99,
-                branch="task/issue-99",
-                actor="agent:run",
-                role="executor",
-                before_state_label="state/run",
-                flow_state=None,
-            )
-
-        mock_block.assert_not_called()
-        store.add_event.assert_called_once()
-        event_args = store.add_event.call_args
-        assert event_args[0][1] == "state_transitioned"
-
-    def test_transition_count_hard_limit_checked_before_state_change(self) -> None:
-        """Hard limit is checked BEFORE state change increment."""
-        store = _make_mock_store()
-        flow_state = {"transition_count": 19}
-
-        with (
-            patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
-            patch(
-                "vibe3.services.issue_failure_service.block_executor_noop_issue"
-            ) as mock_block,
-        ):
-            mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
-                "state/review"
-            )
-            apply_unified_noop_gate(
-                store=store,
-                issue_number=99,
-                branch="task/issue-99",
-                actor="agent:run",
-                role="executor",
-                before_state_label="state/run",
-                flow_state=flow_state,
-            )
-
-        mock_block.assert_called_once()
-        # Event should be transition_count_exceeded, not state_transitioned
-        store.add_event.assert_called_once()
-        event_args = store.add_event.call_args
-        assert event_args[0][1] == "transition_count_exceeded"
-
-    def test_transition_count_defaults_to_zero_when_missing(self) -> None:
-        """transition_count defaults to 0 when missing from flow_state."""
-        store = _make_mock_store()
-        flow_state = {}  # No transition_count key
-
-        with (
-            patch("vibe3.clients.github_client.GitHubClient") as mock_gh,
-            patch(
-                "vibe3.services.issue_failure_service.block_executor_noop_issue"
-            ) as mock_block,
-        ):
-            mock_gh.return_value.view_issue.return_value = _make_github_issue_payload(
-                "state/review"
-            )
-            apply_unified_noop_gate(
-                store=store,
-                issue_number=99,
-                branch="task/issue-99",
-                actor="agent:run",
-                role="executor",
-                before_state_label="state/run",
-                flow_state=flow_state,
-            )
-
-        mock_block.assert_not_called()
-        assert flow_state["transition_count"] == 1  # 0 + 1 = 1
-        store.add_event.assert_called_once()
-        event_args = store.add_event.call_args
-        assert event_args[0][1] == "state_transitioned"
